refactor(scatter-plot): B4 WebGLRenderer god-class decomposition (F-50,F-61,F-29)
Shrink the 2105-line WebGLRenderer to a 1147-line facade (-958) by extracting its
natural seams; strictly behavior-preserving (no sanctioned change).

- F-50: route the remaining framebuffer-teardown triples through B3's existing
  destroyFramebuffer(gl,fb) (no new module).
- F-61: GLResources holder owns the GPU inventory (programs, 6 vertex buffers + quad,
  VAO, label texture, linearFramebuffer) with createAll/validate/deleteAll/reset;
  dirty-flag cache signatures stay on the renderer. validate() is a byte-faithful
  mirror of the original isRendererStateValid (deliberately does NOT check quadBuffer
  or linearFramebuffer — adding those would change when ensureGL resets).
- F-29: split into ContextLossController (single webglcontextlost listener per the
  post-B1 model; markLost idempotent -> onContextLost) and ExportRenderer (offscreen
  subsystem + inset math, consuming the B3 substrate, preserving the F-15 two-pass);
  shared point/gamma shader sources factored into export-shaders.ts. The facade
  preserves every public renderer method signature (the scatter-plot<->renderer contract).

Gates: type-check, core vitest (1144), utils vitest, build, lint (0 err), knip;
e2e dataset-recovery 5/5 + figure-editor 4/4; visual 6/6 (0 drift); F-01 + F-15
oracles both still PASS (split perturbs neither sanctioned behavior).
